backend/requirements.txt
FastAPI
pymongo
motor
uvicorn
python-dotenv
passlib
bcrypt
python-multipart
python-jose
email-validator
PyJWT
python-jose[cryptography]
httpx
requests